OMR AOT: Changes to existing OMR AOT classes
This commit is a final stage to allow previously introduced set of
commits to be included in the build process.
The changes to OMRAheadOfTimeCompile.cpp are inspired by Irwin D'Souza
work on OpenJ9 repository, that enable processing relocation headers
that are platform and architecture independent.
The changes to OMRAheadOfTimeCompiler.hpp are to provide access to the
RelocationRuntime structure that stores the relocation records in the
process of their creation.
The changes to OMRCodeGenerator are addressing creation of AheadOfTime
Compilation object, as well as generation of relocation
records that are OMR specific. Those changes are guarded
by compileRelocatableCode boolean field that is defined in the
OMRCompilation.hpp
Changes to OMRCompilation.hpp involve the pointer to the
RelocationRuntime object to enable access to it.
The changes to FE_base_t.hpp are to allow the allocation of relocation
data that are less than 2048 bytes, because not all of the language
runtimes require such big size of the buffer for relocation records.
The x/OMRAheadOfTimeCompile.hpp and x/OMRAheadOfTimeCompile.cpp are
subclasses that allow architecture specific relocations header
generation similar to the way it is organized in Openj9.

The changes in */CMakeLists.txt are to enable compilation
of the newly added files.

Signed-off-by: Georgiy Krylov <georgiy.krylov@unb.ca>