WIP: Replace placeholder DataTypes enum values
The DataTypes enum contains some placeholders representing values
such as the number of different categories of DataTypes. To prevent
those placeholders from being treated as though they are a value
representing a valid DataType, mark the first and last of each category
of DataTypes instead and remove the placeholders from the enum.
For those placeholder values used extensively elsewhere, replace them
with a constant in the TR namespace of the same value.

Introduce FirstTRScalarType and LastTRScalarType in DataTypesEnum.hpp
to indicate to downstream projects that they should set these values
appropriately based on their extensions to the DataTypes enum.

Issue: #6709