Restore GC Thread Count Limit
Limit the default restore thread count to:  2 X (initial thread count
ceiling to a factor of 8). For example, if the initial thread count was
in a range of 1-8, we would allow up max 16 thread for restore. This
ensures the restore thread count is reasonable considering the GC was
initialized to optimize for the initial startup GC thread count (e.g.,
allocated split lists).

Signed-off-by: Salman Rana <salman.rana@ibm.com>