Change arraycmp to use vector instructions
Modified both arraycmp and arraycmplen evaluators on Power 8 and Power 9 to
take advantage of vector instructions. Power 10 already has an existing vector
implementation of arraycmp and continues to use it.

Added binary encoder tests for the newly used cmpb, vclzd and vclzw
instructions.

Arraycmp and arraycmplen nodes inconsistantly accept both a 32bit and 64bit
value for the length child. The related evaluators on Power were updated to
check the size of the length node in advance to ensure the correct code is
generated.

Signed-off-by: jimmyk <jimmyk@ca.ibm.com>