Add option to control when idiomRecognition opt is applied
This commit adds a new option -Xjit:idiomRecognitionFrequencyThresholdAtWarm=NNN
which is going to be used in the OpenJ9 downstream project.
The idea is to reduce compilation overhead by restricting
the idiomRecognition optimization to only the blocks that
have a frequency larger than the given threshold.
This threshold will take effect only for compilations
performed at warm optimziation level (or lower).
A value of 0 means that idiomRecognition will be applied
to all blocks. A value of 10001 means the idiomRecognition
will not be applied at all, since 10000 is the largest
possible block frequency.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>