Do not continue to merge back cold path if guard2 block has been removed
`changeBranchDestination` removes unreachable blocks after the
removal of the old edge from guard2 to cold2. In rare cases,
guard2Block could become unreachable and removed. Normally, if
all relevant blocks are removed, later `moveBlockAfterDest`
will not cause any issue since it will be just moving around
all unreachable blocks.
However, in an even more rare case, if some of the blocks are
removed and some of them are not (eg, guard2Block and the previous
block of guard2Block are removed but the next block of guard2Block
is still valid), `moveBlockAfterDest` could end up joining an
invalid/removed block with a valid block. Therefore, if guard2Block
is no longer valid, it should not proceed.

Fixes: eclipse-openj9/openj9#18873

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>