Don't map systemcfg on Power in checkpoint mode
On Power `omrtime_current_time_millis` and `omrtime_hires_clock` can
use the timebase register to very quickly get the current time.
This requires mmapping the special `/proc/ppc64/systemcfg` file.

Some security modules (e.g. SELinux) prevent systemcfg from being mapped
by user processes. If we take a checkpoint in an environment that
that allows systemcfg to be mapped it will not be restorable in an
environment that forbids it.

This patch disables the mapping of systemcfg in checkpoint mode, which
will force the omrtime functions to fall back to conventional Linux time
APIs.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>