Handle unsafe loads in trees left behind by Store Sinking
Store Sinking might leave behind trees that contain commoned loads
if necessary to preserve the order of evaluation of operations.  It
is possible for such trees to contain a load of a symbol that had a
value assigned to it in a tree before the current tree, but that was
also moved by Store Sinking.

Typically such a load would end up being removed by Dead Trees
Elimination, but if that optimization fails to run or fails to run to
completion, such a load of an uninitialized symbol might be left behind.
That can be especially problematic if it is used in an indirect
reference.

This change checks for such cases and makes use of
UnsafeSubexpressionRemover to rework such a tree.  It will keep only the
parts that are safe to evaluate and vandalize any loads that are unsafe
to evaluate.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>