Replace arraycopy source and destination with temps
When preparing for null restricted arraycopy transformation,
there are temps created for the source array and the destination
array. Other sub transformations that run after the preparation
and before null-restrcited arraycopy transformation are not aware
of these temps. These temps might not get updated properly.
This change replaces all the commoned source array node and commoned
destination array node with the temps right after the temps are created.

Fixes: eclipse-openj9/openj9#20522
Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>