Fix short arithmetic op nodes being incorrectly marked as cannotOverflow
For arithmetic operations that have a short range constraint, it
is because one or both of the operands are not known and therefore
cannot be determined as cannotOverflow. While the result of the
arithmetic operation would be in a known range of 16bit signed
short values, it may have been achieved by intentionally causing
an overflow.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>