Check if mul can overflow during BNDCHK simplification
For BNDCHK nodes where the first child is a constant and the second
child is a multiplication operation, with the multiplier child being a
constant, it was getting simplified such that the bound value was
being divided by the multiplier. While this is a valid simplification
that eliminates the multiplication operation, it would only be valid
if we could guarantee that the multiplication was not going to
overflow. This change ensures that this simplification is performed
only for mul operation nodes that have been marked as cannotOverflow.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>